Simplify 'with' and rename 'sequence' -> 'do'
Expect the function given to 'with' to return a single script, not a
list; allows using 'andThen' etc. naturally without wrapping such a
composite script into a singleton list (can always use 'do' if desired).

'do' is more concise than 'sequence', especially important since it is
likely to be often nested within 'with'.