Switch back to 'big record' script startup design
New 'Init' name is reminiscent of 'init' in normal Elm programs and would look appropriately weird if some third-party package requested it as an argument (this is good because passing the entire Init record would be terrible for security). And using the record approach makes for nice and concise user code (nice short type signature on the top level 'script' function, and the function itself only has to accept the arguments that it actually wants).